requirements.txt
pymongo>=4.10.1
discord.py>=2.4.0
python-dotenv>=1.0.1
requests>=2.32.3
GitPython>=3.1.44
tqdm>=4.67.1
aiohttp>=3.11.11